refactor: パッケージ構成を変更 close #94
