+++ api/v1_11/src/main/kotlin/dev/s7a/spigot/item/extension.kt
@@ -1,6 +1,6 @@
 @file:JvmName("Item11")
 
-package dev.s7a.spigot.util
+package dev.s7a.spigot.item
 
 import org.bukkit.inventory.meta.ItemMeta
