+++ api/v1_8/src/main/kotlin/dev/s7a/spigot/player/VirtualPlayer.kt
@@ -1,4 +1,4 @@
-package dev.s7a.spigot.util
+package dev.s7a.spigot.player
 
 import org.bukkit.Bukkit
 import org.bukkit.OfflinePlayer
