+++ example/player/src/main/kotlin/dev/s7a/spigot/example/player/Main.kt
@@ -1,8 +1,8 @@
 package dev.s7a.spigot.example.player
 
 import dev.s7a.spigot.listener.registerListener
-import dev.s7a.spigot.util.VirtualPlayer
-import dev.s7a.spigot.util.VirtualPlayer.Companion.toVirtual
+import dev.s7a.spigot.player.VirtualPlayer
+import dev.s7a.spigot.player.VirtualPlayer.Companion.toVirtual
 import org.bukkit.OfflinePlayer
 import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
